pyproject.toml
[project]
name = "boj-jandi-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.128.0",
    "httpx[http2]>=0.28.1",
    "jinja2>=3.1.6",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
